# TRAAN_Integration_Phase5/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIH Disaster Response API",
    "description": "Disaster Response Coordination Platform \u2014 Backend API\n\n**SIH 2026** \u00b7 Component D \u2014 Backend Core\n\nThis Swagger UI is the live contract that Android, Dashboard, and AI/Optimizer teams build against. Do not change endpoint paths, request fields, or response fields without coordinating with all affected teams.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sos/batch": {
      "post": {
        "tags": [
          "sos",
          "sos"
        ],
        "summary": "Upload Sos Batch",
        "operationId": "upload_sos_batch_api_v1_sos_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/sos/{uuid}/status": {
      "get": {
        "tags": [
          "sos",
          "sos"
        ],
        "summary": "Get Sos Status",
        "operationId": "get_sos_status_api_v1_sos__uuid__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SOSStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchResponse": {
        "properties": {
          "accepted_uuids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accepted Uuids"
          },
          "duplicate_uuids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Duplicate Uuids"
          }
        },
        "type": "object",
        "required": [
          "accepted_uuids",
          "duplicate_uuids"
        ],
        "title": "BatchResponse"
      },
      "BatchUpload": {
        "properties": {
          "gateway_device_id": {
            "type": "string",
            "title": "Gateway Device Id"
          },
          "gateway_location": {
            "$ref": "#/components/schemas/app__models__sos__Location"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          },
          "sos_batch": {
            "items": {
              "$ref": "#/components/schemas/SOSRequest"
            },
            "type": "array",
            "title": "Sos Batch"
          }
        },
        "type": "object",
        "required": [
          "gateway_device_id",
          "gateway_location",
          "uploaded_at",
          "sos_batch"
        ],
        "title": "BatchUpload"
      },
...
      "EmergencyType": {
        "type": "string",
        "enum": [
          "medical",
          "trapped",
          "structural_collapse",
          "flood_rescue",
          "fire",
          "missing_person",
          "unspecified"
        ],
        "title": "EmergencyType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SOSRequest": {
        "properties": {
          "uuid": {
            "type": "string",
            "title": "Uuid"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "location": {
            "$ref": "#/components/schemas/app__models__sos__Location"
          },
          "is_quick_sos": {
            "type": "boolean",
            "title": "Is Quick Sos"
          },
          "emergency_type": {
            "$ref": "#/components/schemas/EmergencyType"
          },
          "severity_hint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SeverityHint"
              },
              {
                "type": "null"
              }
            ]
          },
          "people_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "People Count"
          },
          "medical_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medical Snapshot"
          },
          "custom_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Message"
          },
          "contact_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Number"
          },
          "relay_hop_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Relay Hop Count"
          },
          "last_relayed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Relayed At"
          },
          "status": {
            "$ref": "#/components/schemas/SOSStatus"
          }
        },
        "type": "object",
        "required": [
          "uuid",
          "device_id",
          "created_at",
          "location",
          "is_quick_sos",
          "emergency_type",
          "relay_hop_count",
          "last_relayed_at",
          "status"
        ],
        "title": "SOSRequest"
      },
      "SOSStatus": {
        "type": "string",
        "enum": [
          "pending_local",
          "in_relay",
          "uploaded"
        ],
        "title": "SOSStatus"
      },
      "SOSStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SOSStatusResponse"
      },
      "SeverityHint": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "SeverityHint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__models__sos__Location": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng"
          },
          "accuracy_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy M"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "title": "Location"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}